+++ salasaga/ide/functions/text_tags/text_layer_create_font_size_tag.c
@@ -36,7 +36,8 @@
 #include "../../salasaga_types.h"
 #include "../../externs.h"
 
-GtkTextTag *text_layer_create_font_size_tag(gfloat font_size)
+
+GtkTextTag *text_layer_create_font_size_tag(gdouble font_size)
 {
 	// Local variables
 	GtkTextTag			*text_size_text_tag;	// Text size text tag
